继续优化错误: LYSKit/LYSKit/NSDate+LYSCategory.m:75:61: warning: values of type 'NSInteger' should not be used as format arguments; add an explicit cast to 'long' instead [-Wformat]

+++ LYSKit/NSDate+LYSCategory.m
@@ -50,8 +50,8 @@ formatter;\
 + (NSString *)ly_compareCurrentTime:(NSDate *)compareDate
 {
     NSTimeInterval timeInterval = -[compareDate timeIntervalSinceNow];
-    NSInteger time = round(timeInterval);
-    NSInteger temp = 0;
+    double time = round(timeInterval);
+    long temp = 0;
     NSString *result = nil;
     if (time >= 0) {
         if ((temp = time) < 60)
